requirements.txt
# LLM API客户端
openai>=1.0.0
anthropic>=0.7.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 数据处理
pydantic>=2.0.0
typing-extensions>=4.0.0

# 文件处理
pathlib2>=2.3.0

# 测试
pytest>=7.0.0
pytest-cov>=4.0.0

# 代码格式化
black>=23.0.0
isort>=5.12.0

# 类型检查
mypy>=1.0.0

# 日志
loguru>=0.7.0

# 异步支持
aiohttp>=3.8.0
asyncio-throttle>=1.0.0 